refactor:initial PortfolioTabsSection.jsx and TabsSection.jsx
