+++ frontend/src/components/Portfolios/PortfolioTabsSection/PortfolioTabsSection.jsx
@@ -1,7 +1,8 @@
-import styles from "./TabsSection.module.scss";
+import styles from "./PortfolioTabsSection.module.scss";
 import { Link, useLocation } from "react-router-dom";
+import TabsSection from "../../UI/TabsSection";
 
-const TabsSection = ({ portfolioId }) => {
+const PortfolioTabsSection = ({ portfolioId }) => {
     const location = useLocation();
 
     const selected = `font-sans-2xs text-bold ${styles.listItemSelected}`;
@@ -35,15 +36,9 @@ const TabsSection = ({ portfolioId }) => {
 
     return (
         <>
-            <nav
-                className={`margin-bottom-4 ${styles.tabsList}`}
-                aria-label={"Portfolio Tab Sections"}
-                role={"navigation"}
-            >
-                {links}
-            </nav>
+            <TabsSection links={links} label="Portfolio Tabs Section" />
         </>
     );
 };
 
-export default TabsSection;
+export default PortfolioTabsSection;
