+++ frontend/src/components/Portfolios/PortfolioTabsSection/index.js
@@ -1,0 +1,1 @@
+export { default } from "./PortfolioTabsSection";
