+++ frontend/src/components/UI/TabsSection/index.js
@@ -1,0 +1,1 @@
+export { default } from "./TabsSection";
